introducing `qtop` --- like `top(1)` for qorb
This commit adds a really bad implementation of a `qorb` stats tool and
interface for collecting stats from a pool. Stats are collected from
each slot set in the pool and exported over websockets periodically, and
there's a `qtop` binary which serves as a websocket client that displays
those stats.

This is a pretty rough proof of concept, and there's a lot of stuff that
has yet to be done. This includes, but is certainly not limited to:

- [ ] Collect more stats! Error counts, connections/sec, etc would be
  nice...
- [ ] Support stats from multiple `qorb` pools. Right now, there's a
  websocket endpoint that exports stats from a single pool. It would be
  better if the dropshot API could support exporting stats from multiple
  pools, along with (maybe) some way to discover all of the pools that
  there are. This would be nice to figure out later.
- [ ] Optimize the stats collection. For one thing, we could probably
  get rid of the mutexen by using atomic counters. We could probably
  also piggy-back the existing reference counts for the "return
  connection to pool" channel as a way of tracking the number of active
  claims per pool, rather than implementing our own counters --- the
  Tokio MPSC will update these counters anyway, so we may as well use
  them rather than adding an additional layer of counters. And there's
  probably more we can do along those lines...
- [ ] Make the `qtop` TUI not be bad