more idiomatic `tracing`
This commit makes a couple small style tweaks to the use of `tracing` in
`qorb`. In particular, I've changed the use of span targets that include
the fully-qualified name of the instrumented function to span *names*
that include the type on which the associated function is...associated
with, as the target for an idiomatic `tracing` span or event is
generally module-equivalent, while a span _name_ represents the
particular function.

Additionally, I've changed a few events to use field shorthand syntax so
that we don't have to write the field name out twice, which I think is
just a little bit nicer.