Terminate all slots in a set when clients are gone
Currently, the `SetWorker::run` function will terminate all per-slot
tasks in the set when termination is explicitly signaled. However, it
does *not* currently do this when the `SetWorker` shuts down because
there are no more clients, so we don't clean up the per-slot tasks in
that case. I noticed this [while reviewing][1] PR #45, but left the
review only after it merged.

This commit fixes that by moving the termination code out of the loop,
and making both the "explicit termination channel fired" and the "no
more clients" cases `break` out of the loop into the termination code,
instead of `return`ing.

[1]: https://github.com/oxidecomputer/qorb/pull/45#discussion_r1802145192