Make `Pool::new()` fallible
This makes the `Pool` constructor fallible, since it now automatically
registers the USDT probes internally (if the feature is enabled). To let
applications decide how to handle that, we return the pool both in the
`Ok(_)` and `Err(_)` variants of the result, so they can either unwrap
or unwrap the error, and get the pool regardless of whether the
registration worked.